+++ data/poly-perc-demo.pset
@@ -1,0 +1,6 @@
+"clock": 1
+"bpm": 120
+"clock_out": 1
+"EXT": 1
+"ENG": 1
+"midi_out_device": 1
