+++ lib/library.lua
@@ -3,8 +3,9 @@ ops['#'] = include("lib/library/_comment")
 ops['='] = include("lib/library/_osc_out")
 ops['*'] = include("lib/library/_bang")
 ops['?'] = include("lib/library/_levels")
-ops["'"] = include("lib/library/_timber")
-ops['"'] = include("lib/library/_timber_param")
+-- ops["'"] = include("lib/library/_timber")
+-- ops['"'] = include("lib/library/_timber_param")
+ops['|'] = include("lib/library/_polyperc")
 ops[':'] = include("lib/library/_midi")
 ops['&'] = include("lib/library/_midi_in")
 ops['!'] = include("lib/library/_cc")
